+++ src/main/java/fiskfille/tf/common/energon/power/IEnergyContainer.java
@@ -7,7 +7,7 @@ public interface IEnergyContainer {
 
 	float getEnergy();
 
-	float getMaxEnergy();
+	int getMaxEnergy();
 
 	float getEnergyUsage();
 }
