+++ src/main/java/fiskfille/tf/common/energon/power/IEnergyContainerItem.java
@@ -9,5 +9,5 @@ public interface IEnergyContainerItem {
 
 	float getEnergyStored(ItemStack itemstack);
 
-	float getEnergyCapacity(ItemStack itemstack);
+	int getEnergyCapacity(ItemStack itemstack);
 }
